+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
-- New events in core listener for comments (create, edit, delete, mark as "spam", mark as "not spam", trash, untrash, approve, unapprove, duplicate triggered).
-- New events in WPMU listener for users (mark as "spam", mark as "not spam").
+- New events in core listener for posts (trashed, untrashed, drafted, draft saved, published, privately published, scheduled, unscheduled, pending review).
+- New events in core listener for comments (created, edited, deleted, marked as "spam", marked as "not spam", trashed, untrashed, approved, unapproved, duplicate triggered).
+- New events in core listener for menus (created, edited, deleted, item added, item updated).
+- New events in WPMU listener for users (marked as "spam", marked as "not spam").
 - The core listener can now detect plugin and theme installations/updates.
 - The core listener can now detect translations updates.
 ### Changed